drawer + alterando fontes
