--- app/src/main/java/com/e/jarvis/UI/gallery/GalleryFragment.kt
@@ -1,38 +1,0 @@
-package com.e.jarvis.ui.gallery
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
-import androidx.navigation.Navigation
-import com.e.jarvis.R
-import kotlinx.android.synthetic.main.fragment_gallery.view.*
-
-class GalleryFragment : Fragment() {
-
-    private lateinit var galleryViewModel: GalleryViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-//        galleryViewModel =
-//            ViewModelProvider(this).get(GalleryViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_gallery, container, false)
-//        val textView: TextView = root.findViewById(R.id.text_gallery)
-//        galleryViewModel.text.observe(viewLifecycleOwner, Observer {
-//            textView.text = it
-//        })
-
-        root.text_gallery.setOnClickListener {
-            Navigation.findNavController(root).navigate(R.id.action_nav_gallery_to_nav_home)
-        }
-
-        return root
-    }
-}
