+++ app/src/main/java/com/e/jarvis/ui/question/QuestionFragment.kt
@@ -1,0 +1,22 @@
+package com.e.jarvis.ui.question
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.e.jarvis.R
+
+class QuestionFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_question, container, false)
+    }
+
+    companion object {
+        fun newInstance() = QuestionFragment()
+    }
+}
