+++ app/src/main/java/com/e/jarvis/ui/quiz/QuizFragment.kt
@@ -1,22 +1,25 @@
-package com.e.jarvis.UI
+package com.e.jarvis.ui.quiz
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.e.jarvis.R
 
-class QuestionFragment : Fragment() {
+class QuizFragment : Fragment() {
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_question, container, false)
-    }
+        val view: View = inflater.inflate(R.layout.fragment_quiz, container, false)
+        
+        return view
 
-    companion object {
-        fun newInstance() = QuestionFragment()
+    }
+    companion object{
+        fun newInstance() = QuizFragment()
     }
 }
