+++ app/src/main/res/layout/activity_main_drawer.xml
@@ -20,6 +20,11 @@
         android:layout_height="match_parent"
         android:layout_gravity="start"
         android:fitsSystemWindows="true"
+        android:backgroundTint="#080A07"
+        android:background="#080A07"
+        app:itemTextAppearance="@style/fontePrincipal"
+        app:itemTextColor="@color/white"
+        app:itemIconTint="@color/white"
         app:headerLayout="@layout/nav_header_main"
         app:menu="@menu/activity_main_drawer" />
