+++ app/src/main/res/layout/fragment_home.xml
@@ -38,28 +38,30 @@
 
         <TextView
             android:id="@+id/home_mais_procurados"
+            style="@style/fontePrincipal"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="124dp"
-            style="@style/norwester"
+            android:layout_marginTop="36dp"
             android:text="@string/top_personagens"
             android:textColor="@color/white"
             android:textSize="20sp"
             android:textStyle="bold"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.497"
+            app:layout_constraintHorizontal_bias="0.496"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toBottomOf="@+id/barra_pesquisa1" />
 
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="450dp"
+            android:layout_marginTop="24dp"
             android:gravity="center"
             android:orientation="vertical"
             android:paddingRight="15dp"
-            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent">
+            app:layout_constraintHorizontal_bias="0.0"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/home_mais_procurados">
 
             <LinearLayout
                 android:layout_width="wrap_content"
