+++ app/src/main/res/layout/fragment_pesquisa.xml
@@ -11,7 +11,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="400dp"
-        android:background="@drawable/search_bw">
+        >
 
         <SearchView
             android:id="@+id/barra_pesquisa2"
