+++ app/src/main/res/layout/fragment_question.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.QuestionFragment">
+    tools:context=".ui.question.QuestionFragment">
     <ImageView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
