+++ app/src/main/res/layout/fragment_quiz.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.QuizFragment">
+    tools:context=".ui.quiz.QuizFragment">
 
     <ImageView
         android:layout_width="match_parent"
