--- app/src/main/res/layout/side_menu.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:orientation="vertical"
-    android:backgroundTint="#080A07"
-    android:background="#080A07">
-    <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/marvellogo"
-        />
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rcDishes"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:listitem="@layout/item_menu"
-         />
-</LinearLayout>
