+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -1,20 +1,27 @@
 <?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
+<menu
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     tools:showIn="navigation_view">
-
-    <group android:checkableBehavior="single">
+    <group
+        android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_slideshow"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="@string/menu_home" />
-        <item
-            android:id="@+id/nav_gallery"
-            android:icon="@drawable/ic_menu_gallery"
-            android:title="@string/menu_gallery" />
+            android:id="@+id/loginFragment"
+            android:icon="@drawable/ic_baseline_account_circle_24"
+            android:title="@string/Login"
+            />
         <item
             android:id="@+id/nav_home"
-            android:icon="@drawable/ic_menu_slideshow"
-            android:title="@string/menu_slideshow" />
+            android:icon="@drawable/ic_baseline_home_24"
+            android:title="@string/Home"
+            />
+        <item
+            android:id="@+id/quizFragment"
+            android:icon="@drawable/ic_baseline_view_carousel_24"
+            android:title="@string/Quiz" />
+        <item
+            android:id="@+id/favoritosPersonagemFragment"
+            android:icon="@drawable/ic_baseline_star_border_24"
+            android:title="@string/Favoritos" />
     </group>
 </menu>
