+++ app/src/main/res/values/strings.xml
@@ -13,33 +13,27 @@
     <string name="button_google">LOGIN WITH GOOGLE</string>
 
     <!--Tela de Cadastro-->
-
     <string name="prompt_confirm_password">Confirm password</string>
     <string name="prompt_birthday">Birthday</string>
     <string name="prompt_register">REGISTER USER</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
 
-
+    <!--Tela de procura / Main -->
     <string name="nome_personagem">Wolverine</string>
-    <string name="personagens_encontrados">Personagens encontrados...</string>
-    <string name="top_personagens">Personagens mais procurados...</string>
+    <string name="personagens_encontrados">All find results...</string>
+    <string name="top_personagens">Top character...</string>
     <string name="QuizIntro">teste</string>
 
-    <string name="seus_favoritos">Seus favoritos</string>
+    <!--Tela de Favoritos -->
+    <string name="seus_favoritos">Your Favorites</string>
     <string name="nome_hq">X-FORCE</string>
 
-    <string name="title_activity_main_drawer">MainActivityDrawer</string>
-    <string name="navigation_drawer_open">Open navigation drawer</string>
-    <string name="navigation_drawer_close">Close navigation drawer</string>
-    <string name="nav_header_title">Android Studio</string>
-    <string name="nav_header_subtitle">android.studio@android.com</string>
-    <string name="nav_header_desc">Navigation header</string>
-    <string name="action_settings">Settings</string>
-
+    <!--Drawer -->
     <string name="menu_home">Home</string>
-    <string name="menu_gallery">Gallery</string>
-    <string name="menu_slideshow">Slideshow</string>
+    <string name="Login">Login</string>
+    <string name="Quiz">Quiz</string>
+    <string name="Favoritos">Favorites</string>
+    <string name="Home">Home</string>
+    <string name="Or">Or</string>
 
 
 </resources>
